the great un-asyncing
It was proving really tricky to get AI working from within lua.
Especially with tools and JSON conversion.

A lot of this was traced back to problems supporting async.

The cause for us to use async was primarily that the AI library uses async.
So instead of letting that work its way through all our code, we confine it
by doing all AI stuff on its own threads, where it can be as async as it likes,
and our non-async code blocking on ai calls using channels.

This is a bit ugly to set up, but now that the bridge object is sync again,
all the interfacing with WASM and luau is much nicer.

Now we can get back to adding the features that were blocked on this.. Yay!